+++ src/components/Settings/screens/SettingsFiltersScreen.tsx
@@ -1,0 +1,27 @@
+import React from 'react';
+import { INavigationProps } from '@src/types';
+import { ScrollView } from 'tamagui';
+import Table from '@components/Common/Table/Table';
+
+export default function SettingsFiltersScreen({
+  navigation,
+}: INavigationProps): React.JSX.Element {
+  return (
+    <ScrollView flex={1}>
+      <Table.Container>
+        <Table.Section header="Filters">
+          <Table.Cell
+            label="Instance Filters"
+            onPress={() => navigation.navigate('Filter', { type: 'instance' })}
+            useChevron
+          />
+          <Table.Cell
+            label="Keyword Filters"
+            onPress={() => navigation.navigate('Filter', { type: 'keyword' })}
+            useChevron
+          />
+        </Table.Section>
+      </Table.Container>
+    </ScrollView>
+  );
+}