+++ src/state/filters/actions/index.ts
@@ -1,0 +1,2 @@
+export * from './keywordFilterActions';
+export * from './instanceFilterActions';