+++ src/state/index.ts
@@ -11,3 +11,4 @@ export * from './post';
 export * from './profile';
 export * from './settings';
 export * from './inbox';
+export * from './filters';